Fix C2220 warning '=='signed/unsigned mismatch (#271)
<!--
Copyright (C) 2023 - 2025 Advanced Micro Devices, Inc. All rights
reserved.
Licensed under the MIT License.
-->
# Summary of Changes

[opensource] Fix error C2220: the following warning is treated as an
error：
```
 C:\Users\Administrator\opensource\source\MorphiZen\vaip-core\src\file_stream.cpp(92,3):
          see reference to function template instantiation 'std::string *google::Check_EQImpl<unsigned __int64,__int64>(const T1 &,const T2 &,const char *)' being compiled
          with
          [
              T1=unsigned __int64,
              T2=__int64
          ]
 ```

+++ vaip-core/src/file_stream.cpp
@@ -86,7 +86,7 @@ std::streambuf::int_type FileBuf::overflow(int_type ch) {
       });
   auto r = std::fseek(file_, (long)put_pos_, SEEK_SET);
   CHECK(r == 0) << " conner case: fseek fail";
-  auto num_of_elements = end - pbase();
+  auto num_of_elements = static_cast<std::size_t>(end - pbase());
   auto num_of_element_written =
       std::fwrite(pbase(), sizeof(char_type), num_of_elements, file_);
   CHECK_EQ(num_of_element_written, num_of_elements) << " error writing.";
@@ -206,7 +206,7 @@ bool FileBuf::flush_buffer() {
   std::ptrdiff_t count = pptr() - pbase();
   if (count > 0) {
     std::size_t written = std::fwrite(pbase(), sizeof(char_type), count, file_);
-    CHECK_EQ(written, count);
+    CHECK_EQ(written, static_cast<std::size_t>(count));
     setp(put_buffer_.data(),
          put_buffer_.data() + put_buffer_.size() -
              1 /*overflow() needs one more ch to be written*/);
